Create UserActions modal, delete GetRegister page
