--- client/src/pages/GetStartedPage/getstarted.css
@@ -1,58 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-}
-
-.item0 {
-  flex: auto;
-  background-color: #0000;
-  color: #F6F6F3;
-}
-
-/* create an account */
-.create {
-  font-family: 'Prata', serif;
-  color: #000;
-  text-align: center;
-}
-
-/* create an account description */
-.text1 {
- font-family: 'Slabo 27px', serif;
- color: #000;
-}
-
-/* reasons to create an account  */
-.reasons {
- color: #000;
-}
-
-.item1 {
-  flex: auto;
-  font-size: 30px;
-  background-color: #0000;
-  border: 3px solid #DBD9DD;
-  color: #E6E6E8;
-  margin-top: 50px; 
-  margin-bottom: 50px; 
-  max-width: 415px; 
-  /* margin-left: 455px;  */
-  background-color: #F7F3F4;
-  
-}
-
-/* user's input email and password */
-.form-control {
-  font-family: 'Slabo 27px', serif;
-  padding: 22px 3px;
-  margin-left: 9px;
-  width: 95%;
-}
-
-/* "Email address" and "password" txt */
-.control-label {
-  font-family: 'Prata', serif;
-  margin-left: 9px;
-  color: #000;
-}
