--- client/src/pages/GetStartedPage/index.js
@@ -1,1 +1,0 @@
-export {default} from './getstarted.js';
